Use float for texture coordinates

+++ src/core/qgspainting.h
@@ -94,9 +94,9 @@ class CORE_EXPORT QgsPainting
       QPainter *painter,
       const QPolygonF &triangle,
       const QImage &textureImage,
-      double textureX1, double textureY1,
-      double textureX2, double textureY2,
-      double textureX3, double textureY3
+      float textureX1, float textureY1,
+      float textureX2, float textureY2,
+      float textureX3, float textureY3
     );
 
 };
